# docs/swagger.yaml
basePath: /
definitions:
  CurrencyControl_internal_delivery_dto.ApprovalItemResponse:
    properties:
      amount:
        type: number
      approval_status:
        type: string
      branch_id:
        type: integer
      branch_name:
        type: string
      compliance_comment:
        type: string
      compliance_decision:
        type: string
      compliance_reviewed_at:
        type: string
      compliance_reviewed_by:
        type: string
      counterparty_name:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      currency:
        type: string
      currency_control_comment:
        type: string
      currency_control_decision:
        type: string
      currency_control_reviewed_at:
        type: string
      currency_control_reviewed_by:
        type: string
      document_date:
        type: string
      document_number:
        type: string
      entity_id:
        type: integer
      entity_type:
        type: string
      rejection_reason:
        type: string
      subject:
        type: string
    type: object
...
  CurrencyControl_internal_delivery_dto.ComplianceDecisionRequest:
    properties:
      comment:
        example: Причина отказа обязательна при выборе reject
        type: string
      decision:
        description: '"approve" (Одобрено), "reject" (Отказ)'
        enum:
        - approve
        - reject
        example: approve
        type: string
    type: object
...
  CurrencyControl_internal_delivery_dto.CurrencyControlDecisionRequest:
    properties:
      comment:
        example: Предмет контракта проверен
        type: string
      decision:
        description: '"accepted" (Принято), "revision" (На доработку), "rejected"
          (Отклонено)'
        enum:
        - accepted
        - revision
        - rejected
        example: accepted
        type: string
    type: object
...
  CurrencyControl_internal_delivery_dto.PendingApprovalsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/CurrencyControl_internal_delivery_dto.ApprovalItemResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  delivery_http.CommonError:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: API сервера для системы валютного контроля.
  title: Currency Control API
  version: "1.0"
paths:
...
  /api/approvals/{entity_type}/{id}:
    get:
      description: Возвращает текущий статус двухэтапной проверки документа, историю
        решений и комментарии.
      parameters:
      - description: Тип документа (contract, invoice, gtd, additional_agreement)
        in: path
        name: entity_type
        required: true
        type: string
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CurrencyControl_internal_delivery_dto.ApprovalItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Детальная информация о статусе авторизации документа
      tags:
      - Approvals
  /api/approvals/{entity_type}/{id}/compliance:
    post:
      consumes:
      - application/json
      description: Сотрудник комплаенса может либо одобрить документ ("approve"),
        либо отклонить ("reject") с обязательным указанием причины в комментарии.
      parameters:
      - description: Тип документа (contract, invoice, gtd, additional_agreement)
        in: path
        name: entity_type
        required: true
        type: string
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      - description: Решение комплаенс-контроля
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CurrencyControl_internal_delivery_dto.ComplianceDecisionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CurrencyControl_internal_delivery_dto.ApprovalItemResponse'
        "400":
          description: Причина отказа обязательна при отклонении
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Проверка документа отделом комплаенс-контроля
      tags:
      - Approvals
  /api/approvals/{entity_type}/{id}/currency-control:
    post:
      consumes:
      - application/json
      description: 'Сотрудник валютного контроля изучает предмет документа и выносит
        одно из трех решений: "accepted" (Принято), "revision" (На доработку), "rejected"
        (Отклонено).'
      parameters:
      - description: Тип документа (contract, invoice, gtd, additional_agreement)
        in: path
        name: entity_type
        required: true
        type: string
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      - description: Решение валютного контроля
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CurrencyControl_internal_delivery_dto.CurrencyControlDecisionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CurrencyControl_internal_delivery_dto.ApprovalItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Проверка документа отделом валютного контроля
      tags:
      - Approvals
  /api/approvals/pending:
    get:
      description: Возвращает документы, ожидающие проверки валютным контролем или
        комплаенсом, либо возвращенные на доработку.
      parameters:
      - description: 'Этап согласования: currency_control, compliance, revision'
        in: query
        name: stage
        type: string
      - description: 'Тип документа: contract, invoice, gtd, additional_agreement'
        in: query
        name: entity_type
        type: string
      - description: ID филиала (для фильтрации)
        in: query
        name: branch_id
        type: integer
      - description: Номер страницы (по умолчанию 1)
        in: query
        name: page
        type: integer
      - description: Размер страницы (по умолчанию 20, макс 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CurrencyControl_internal_delivery_dto.PendingApprovalsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Список документов на согласовании
      tags:
      - Approvals
...
securityDefinitions:
  ApiKeyAuth:
    description: Введите Bearer access_token (полученный при /auth/login или /auth/refresh)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"